Moved "Person" interface into the Briar Patch
